[FE] Re-route Header Navbar, Create Page Video Details, Rename Bottom Footer
