+++ src/app/components/navigations/HeaderNavbarSVP.tsx
@@ -1,5 +1,5 @@
 "use client";
-import { useEffect, useRef, useState } from "react";
+import React, { useEffect, useRef, useState } from "react";
 import Image from "next/image";
 import Link from "next/link";
 import { usePathname } from "next/navigation";
@@ -31,6 +31,11 @@ export default function HeaderNavbarSVP({
   const [isActionsOpened, setIsActionsOpened] = useState(false);
   const wrapperRef = useRef<HTMLDivElement>(null);
   const pathname = usePathname();
+
+  // Path yang tidak mau menampilkan Navbar & Footer
+  const disallowedPath = ["/auth", "/checkout", "/event"];
+  const isDisallowedPage =
+    pathname === "/" || disallowedPath.some((path) => pathname.includes(path));
 
   // --- Detect screen size
   // useEffect(() => {
@@ -78,75 +83,79 @@ export default function HeaderNavbarSVP({
   };
 
   return (
-    <div className="navbar-root flex sticky w-full bg-white top-0 left-0 items-center justify-center shadow-md z-[90]">
-      <div className="navbar-container flex items-center w-full justify-between py-3 px-5 lg:px-0 lg:py-4 lg:max-w-[960px] xl:max-w-[1208px]">
-        <Link href={"/"}>
-          <Image
-            className="max-w-[180px]"
-            src={
-              "https://tskubmriuclmbcfmaiur.supabase.co/storage/v1/object/public/sevenpreneur//logo-sevenpreneur-main.svg"
-            }
-            alt="Homepage"
-            width={320}
-            height={320}
-          />
-        </Link>
-        {isLoggedIn ? (
-          <div
-            className="user-menu relative flex"
-            ref={wrapperRef}
-            onClick={handleActionsDropdown}
-          >
-            <AvatarBadgeSVP userAvatar={userAvatar} />
-            <AppDropdown
-              isOpen={isActionsOpened}
-              onClose={() => setIsActionsOpened(false)}
-              alignMobile="right"
-            >
-              {/* <Link href={`https://www.${domain}/account`}>
+    <React.Fragment>
+      {!isDisallowedPage && (
+        <div className="navbar-root flex sticky w-full bg-white top-0 left-0 items-center justify-center shadow-md z-[90]">
+          <div className="navbar-container flex items-center w-full justify-between py-3 px-5 lg:px-0 lg:py-4 lg:max-w-[960px] xl:max-w-[1208px]">
+            <Link href={"/"}>
+              <Image
+                className="max-w-[180px]"
+                src={
+                  "https://tskubmriuclmbcfmaiur.supabase.co/storage/v1/object/public/sevenpreneur//logo-sevenpreneur-main.svg"
+                }
+                alt="Homepage"
+                width={320}
+                height={320}
+              />
+            </Link>
+            {isLoggedIn ? (
+              <div
+                className="user-menu relative flex"
+                ref={wrapperRef}
+                onClick={handleActionsDropdown}
+              >
+                <AvatarBadgeSVP userAvatar={userAvatar} />
+                <AppDropdown
+                  isOpen={isActionsOpened}
+                  onClose={() => setIsActionsOpened(false)}
+                  alignMobile="right"
+                >
+                  {/* <Link href={`https://www.${domain}/account`}>
                 <AppDropdownItemList
                   menuIcon={<UserRound className="size-4" />}
                   menuName="Profile"
                 />
               </Link> */}
-              {userRole !== 3 && (
-                <Link href={`https://admin.${domain}`}>
-                  <AppDropdownItemList
-                    menuIcon={<Blocks className="size-4" />}
-                    menuName="Dashboard Admin"
-                  />
-                </Link>
-              )}
-              {/* <Link href={`https://agora.${domain}`}>
+                  {userRole !== 3 && (
+                    <Link href={`https://admin.${domain}`}>
+                      <AppDropdownItemList
+                        menuIcon={<Blocks className="size-4" />}
+                        menuName="Dashboard Admin"
+                      />
+                    </Link>
+                  )}
+                  {/* <Link href={`https://agora.${domain}`}>
                 <AppDropdownItemList
                   menuIcon={<BookMarked className="size-4" />}
                   menuName="Agora Learning"
                 />
               </Link> */}
-              <Link href={`https://www.${domain}/transactions`}>
-                <AppDropdownItemList
-                  menuIcon={<Wallet className="size-4" />}
-                  menuName="Transaction"
-                />
+                  <Link href={`https://www.${domain}/transactions`}>
+                    <AppDropdownItemList
+                      menuIcon={<Wallet className="size-4" />}
+                      menuName="Transaction"
+                    />
+                  </Link>
+                  <hr className="my-1" />
+                  <AppDropdownItemList
+                    menuIcon={<LogOut className="size-4" />}
+                    menuName="Sign out"
+                    isDestructive
+                    onClick={handleSignOut}
+                  />
+                </AppDropdown>
+              </div>
+            ) : (
+              <Link href={`/auth/login?redirectTo=${pathname}`}>
+                <AppButton variant="dark" size="defaultRounded">
+                  <UserCircle2 className="size-5" />
+                  Login
+                </AppButton>
               </Link>
-              <hr className="my-1" />
-              <AppDropdownItemList
-                menuIcon={<LogOut className="size-4" />}
-                menuName="Sign out"
-                isDestructive
-                onClick={handleSignOut}
-              />
-            </AppDropdown>
+            )}
           </div>
-        ) : (
-          <Link href={`/auth/login?redirectTo=${pathname}`}>
-            <AppButton variant="dark" size="defaultRounded">
-              <UserCircle2 className="size-5" />
-              Login
-            </AppButton>
-          </Link>
-        )}
-      </div>
-    </div>
+        </div>
+      )}
+    </React.Fragment>
   );
 }
